feat: 일마감 및 인센티브 계산 로직 구현
- DailyClose Entity/Repository/Service/Controller 생성
- ProdIncentive 일마감 시 자동 계산
- incentiveRate int 타입으로 변경
- 인센티브 계산 테스트 완료

+++ src/main/java/com/mes/mes_project/dto/ItemResponseDto.java
@@ -19,7 +19,7 @@ public class ItemResponseDto {
     private String spec; // 스펙
     private String unit; // 단위
     private int unitPrice; //단가
-    private Double incentiveRate; // 인센티브
+    private int incentiveRate; // 인센티브
     private String useYn = "Y";  // 기본값 Y
 
     private String createdBy;   // 생성자
